requirements.txt
python-telegram-bot>=21.0
telethon>=1.35.0
mutagen>=1.45.0
pydub>=0.25.0
ffmpeg-python>=0.2.0
Pillow>=10.0.0
aiofiles>=23.0.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
psutil>=5.9.0